pyproject.toml
[project]
name = "ncu-regulation-kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "langchain>=1.2.15",
    "langchain-core>=1.2.31",
    "langchain-huggingface>=1.2.1",
    "neo4j>=6.1.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "sentencepiece>=0.2.1",
    "torch>=2.5.1",
    "transformers>=5.5.4",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
